Fix issue with ACR private endpoint policy
This commit fixes an issue that prevented ACRs from being deployed with
the policy assigned. While the previous versiond worked properly for
existing ACRs, it prevented new ACRs from being created. The reason is
because of how the ACR deployment template is processed. It does not
include the private endpoint in the template.